server/requirements.txt
acme==2.11.0
amqp==5.2.0
aniso8601==8.0.0
annotated-types==0.6.0
argcomplete==1.10.3
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.8.2
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
cachetools==5.3.3
celery==5.4.0
certbot==2.11.0
certbot-nginx==2.11.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
compressed_rtf==1.0.6
ConfigArgParse==1.7
configobj==5.0.8
cors==1.0.1
cryptography==42.0.5
Cython==3.0.10
Deprecated==1.2.14
distro==1.9.0
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.2.0
extract-msg==0.28.7
filelock==3.13.3
flasgger==0.9.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
future==1.0.0
fuzzywuzzy==0.18.0
gevent==24.2.1
google==3.0.0
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.6
IMAPClient==2.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
josepy==1.14.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
Levenshtein==0.25.1
limits==3.13.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mistune==3.0.2
msgspec==0.18.6
mysqlclient==2.2.4
olefile==0.47
ordered-set==4.1.0
packaging==24.0
parsedatetime==2.6
pdfminer==20191125
pdfminer.six==20191110
pep8==1.7.1
pillow==10.3.0
pluggy==1.4.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybreaker==1.2.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyRFC3339==1.1
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-pptx==0.6.23
pytz==2020.1
PyYAML==6.0.1
rapidfuzz==3.9.0
redis==5.0.5
referencing==0.34.0
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
setuptools==70.0.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.29
textract==1.6.5
tldextract==5.1.2
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.16.0
WTForms==3.1.2
xlrd==1.2.0
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==6.2
